+++ src/db/dexie/index.ts
@@ -1,0 +1,5 @@
+export * from "./db";
+export * from "./DexieStorageAdapter";
+export * from "./mappers";
+export * from "./repositories";
+export * from "./schema";