+++ src/db/dexie/repositories/index.ts
@@ -1,0 +1,6 @@
+export * from "./DexieDailyCheckinsRepository";
+export * from "./DexieJournalRepository";
+export * from "./DexieKnowledgeRepository";
+export * from "./DexieProjectsRepository";
+export * from "./DexieSettingsRepository";
+export * from "./DexieTasksRepository";